Refactor: consolidate views into feature files
Move AddKeychainItemView and KeychainUIView into their respective feature
files for better code organization. This follows a common TCA pattern of
keeping related views and features together.
